Add component identifier to IdentifyTransformExecutionProgressDetails
